+++ example_test.go
@@ -8,31 +8,12 @@ import (
 	"github.com/xybor-x/xylog"
 )
 
-// NOTE: In example_test.go, xylog.StdoutEmitter is not accepted as the standard
-// output. For this reason, in all examples, we must create a new one.
-// In reality, you should use xylog.StdoutEmitter or xylog.StderrEmitter
-// instead.
-
-func Example() {
-	// You can directly use xylog functions to log with the root logger.
+func ExampleLogger() {
+	var emitter = xylog.NewStreamEmitter(os.Stdout)
 	var handler = xylog.GetHandler("")
-	handler.AddEmitter(xylog.NewStreamEmitter(os.Stdout))
-	xylog.AddHandler(handler)
-	xylog.SetLevel(xylog.DEBUG)
+	handler.AddEmitter(emitter)
 
-	xylog.Debug("foo")
-	xylog.Debugf("foo %s", "bar")
-
-	// Output:
-	// foo
-	// foo bar
-}
-
-func ExampleGetLogger() {
-	var handler = xylog.GetHandler("")
-	handler.AddEmitter(xylog.NewStreamEmitter(os.Stdout))
-
-	var logger = xylog.GetLogger("example.GetLogger")
+	var logger = xylog.GetLogger("example.Logger")
 	logger.AddHandler(handler)
 	logger.SetLevel(xylog.DEBUG)
 	logger.Debugf("foo %s", "bar")
@@ -42,8 +23,8 @@ func ExampleGetLogger() {
 }
 
 func ExampleHandler() {
-	// You can use a handler throughout program without storing it in global
-	// scope. All handlers can be identified by their names.
+	// You can use a Handler throughout program because all Handlers can be
+	// identified by their names.
 	var handlerA = xylog.GetHandler("example.Handler")
 	var handlerB = xylog.GetHandler("example.Handler")
 	if handlerA == handlerB {
@@ -52,7 +33,8 @@ func ExampleHandler() {
 		fmt.Println("handlerA != handlerB")
 	}
 
-	// In case name is an empty string, it totally is a fresh handler.
+	// In case the name is an empty string, it creates different Handlers every
+	// call.
 	var handlerC = xylog.GetHandler("")
 	var handlerD = xylog.GetHandler("")
 	if handlerC == handlerD {
@@ -66,11 +48,32 @@ func ExampleHandler() {
 	// handlerC != handlerD
 }
 
-func ExampleFormatter() {
+func ExampleEventLogger() {
+	var emitter = xylog.NewStreamEmitter(os.Stdout)
 	var handler = xylog.GetHandler("")
-	handler.AddEmitter(xylog.NewStreamEmitter(os.Stdout))
-	handler.SetFormatter(xylog.NewTextFormatter(
-		"module=%(name)s level=%(levelname)s %(message)s custom=%(custom)s"))
+	handler.AddEmitter(emitter)
+
+	var logger = xylog.GetLogger("example.EventLogger")
+	logger.AddHandler(handler)
+	logger.SetLevel(xylog.DEBUG)
+	logger.AddField("boss", "foo")
+
+	logger.Event("create").Field("product", 1235).Debug()
+	logger.Event("use").Field("product", "bar").JSON().Debug()
+
+	// Output:
+	// boss=foo event=create product=1235
+	// {"boss":"foo","event":"use","product":"bar"}
+}
+
+func ExampleTextFormatter() {
+	var emitter = xylog.NewStreamEmitter(os.Stdout)
+	var formatter = xylog.NewTextFormatter(
+		"module=%(name)s level=%(levelname)s %(message)s custom=%(custom)s")
+
+	var handler = xylog.GetHandler("")
+	handler.AddEmitter(emitter)
+	handler.SetFormatter(formatter)
 
 	var logger = xylog.GetLogger("example.Formatter")
 	logger.AddExtra("custom", "something")
@@ -82,32 +85,14 @@ func ExampleFormatter() {
 	// module=example.Formatter level=DEBUG product=1235 custom=something
 }
 
-func ExampleEventLogger() {
-	var handler = xylog.GetHandler("")
-	handler.AddEmitter(xylog.NewStreamEmitter(os.Stdout))
-	handler.SetFormatter(xylog.NewTextFormatter(
-		"module=%(name)s level=%(levelname)s %(message)s"))
-
-	var logger = xylog.GetLogger("example.EventLogger")
-	logger.AddHandler(handler)
-	logger.SetLevel(xylog.DEBUG)
-	logger.AddField("boss", "foo")
-
-	logger.Event("create").Field("product", 1235).Debug()
-	logger.Event("use").Field("product", "bar").JSON().Debug()
-
-	// Output:
-	// module=example.EventLogger level=DEBUG boss=foo event=create product=1235
-	// module=example.EventLogger level=DEBUG {"boss":"foo","event":"use","product":"bar"}
-}
-
 func ExampleJSONFormatter() {
+	var emitter = xylog.NewStreamEmitter(os.Stdout)
 	var formatter = xylog.NewJSONFormatter().
 		AddField("module", "name").
 		AddField("level", "levelname").
 		AddField("", "message")
 	var handler = xylog.GetHandler("")
-	handler.AddEmitter(xylog.NewStreamEmitter(os.Stdout))
+	handler.AddEmitter(emitter)
 	handler.SetFormatter(formatter)
 
 	var logger = xylog.GetLogger("example.JSONFormatter")
@@ -119,22 +104,49 @@ func ExampleJSONFormatter() {
 	// {"event":"create","level":"DEBUG","module":"example.JSONFormatter","product":1235}
 }
 
-func ExampleStructureFormatter() {
-	var formatter = xylog.NewStructureFormatter().
+func ExampleStructuredFormatter() {
+	var emitter = xylog.NewStreamEmitter(os.Stdout)
+	var formatter = xylog.NewStructuredFormatter().
 		AddField("module", "name").
 		AddField("level", "levelname").
 		AddField("", "message")
 	var handler = xylog.GetHandler("")
-	handler.AddEmitter(xylog.NewStreamEmitter(os.Stdout))
+	handler.AddEmitter(emitter)
 	handler.SetFormatter(formatter)
 
-	var logger = xylog.GetLogger("example.StructureFormatter")
+	var logger = xylog.GetLogger("example.StructuredFormatter")
 	logger.AddHandler(handler)
 	logger.SetLevel(xylog.DEBUG)
 	logger.Event("create").Field("employee", "david").Debug()
 
 	// Output:
-	// module=example.StructureFormatter level=DEBUG event=create employee=david
+	// module=example.StructuredFormatter level=DEBUG event=create employee=david
+}
+
+// LoggerNameFilter only logs out records belongs to a specified logger.
+type LoggerNameFilter struct {
+	name string
+}
+
+func (f *LoggerNameFilter) Filter(r xylog.LogRecord) bool {
+	return f.name == r.Name
+}
+
+func ExampleFilter() {
+	var emitter = xylog.NewStreamEmitter(os.Stdout)
+	var handler = xylog.GetHandler("")
+	handler.AddEmitter(emitter)
+	handler.AddFilter(&LoggerNameFilter{"example.filter.chat"})
+
+	var logger = xylog.GetLogger("example.filter")
+	logger.AddHandler(handler)
+	logger.SetLevel(xylog.DEBUG)
+
+	xylog.GetLogger("example.filter.auth").Debug("auth foo")
+	xylog.GetLogger("example.filter.chat").Debug("chat foo")
+
+	// Output:
+	// chat foo
 }
 
 func ExampleNewSizeRotatingFileEmitter() {
@@ -143,7 +155,7 @@ func ExampleNewSizeRotatingFileEmitter() {
 	var emitter = xylog.NewSizeRotatingFileEmitter("exampleSize.log", 30, 2)
 	var handler = xylog.GetHandler("")
 	handler.AddEmitter(emitter)
-	handler.SetFormatter(xylog.NewTextFormatter("%(message)s"))
+
 	var logger = xylog.GetLogger("example.SizeRotatingFileEmitter")
 	logger.SetLevel(xylog.DEBUG)
 	logger.AddHandler(handler)
@@ -178,7 +190,6 @@ func ExampleNewTimeRotatingFileEmitter() {
 		"exampleTime.log", time.Second, 2)
 	var handler = xylog.GetHandler("")
 	handler.AddEmitter(emitter)
-	handler.SetFormatter(xylog.NewTextFormatter("%(message)s"))
 	var logger = xylog.GetLogger("example.TimeRotatingFileEmitter")
 	logger.SetLevel(xylog.DEBUG)
 	logger.AddHandler(handler)